+++ jest.setup.js
@@ -1,0 +1,7 @@
+// Jest setup file to make global functions available
+
+// Make fail function available globally
+/* global global */
+global.fail = (message) => {
+  throw new Error(message);
+};